tests: Add first integration tests
We have to exclude the activation of the openvpn client and execution of
the corresponding handlers because there are not valid connection keys
to create a vpn tunnel. We have to exclude tasks like described below.

First define(the value doesn't matter  a variable like this:
    k1599_integration_test: true

and mark your task like this:
    when: k1599_integration_test is not defined

There is a issue with the activation of clamav-daemon. The service can't
be activated inside the docker hosts so it will changed all the time.
Maybe later we could apply a fix for that.

TODO:
 - Move the clamav installation into an own task.
 - Check activation of the clamav service.